Final merge of VC4 mesa's with Vulkan source

+++ mesa 3D driver/src/gallium/frontends/nine/nine_debug.c
@@ -74,7 +74,7 @@ _nine_debug_printf( unsigned long flag,
 
 #if defined(HAVE_PTHREAD)
     if (dbg_flags & DBG_TID)
-        tid = (unsigned long)pthread_self();
+        tid = pthread_self();
 #endif
 
     if (dbg_flags & flag) {
